requirements.txt
python-dotenv
python-telegram-bot
supabase
PyCharacterAI